+++ docs/changes.rst
@@ -39,6 +39,14 @@ Released: not yet
   on differences between pywbem 1.0.0 and earlier versions of python.
   (issue #1405)
 
+* Refactor pywbemcli tests to avoid hiding the default connection file and
+  the names for the connection file and mock cache by moving as part of tests.
+  EnvVar PYWBEMCLI_ALT_HOME_DIR defines alternate directory for connection file
+  and mockcache and is set for all pywbemcli tests. The definition of file
+  names for default connection file and mock cache managed by pywbemcli moved
+  from  pywbemtools/pywbemcli/_utils.py to pywbemcli/_connection_file_names.py.
+  (issue # 1423)
+
 **Known issues:**
 
 * See `list of open issues`_.